datatype: address pre-push review hardening findings
A review of the improved pack/unpack branch ahead of pushing turned up a
set of small correctness-hardening and hygiene issues. None change the
behavior of the paths exercised today; each removes a latent trap or
tightens an invariant so a future edit fails loudly instead of silently.

- convertor: assert that a stack-copying clone of a receive convertor
  never carries a nonzero partial_length, and document why the clone
  does not copy it (send convertors and repositioning callers are safe).
- convertor: expand the comments explaining why a heterogeneous
  convertor may reuse the optimized description unless RESTRICTED, and
  why accelerator repositioning takes the generic stack walk.
- optimizer: add defensive bounds assertions before every descriptor
  emit site, and document that the count-boundary check runs exactly
  once per commit on the final optimized descriptor.
- pack/unpack: drive the heterogeneous partial-blocklen COUNT decrement
  and block-completion test off the mover's returned element count
  rather than the requested count, keeping COUNT, the return value, and
  the pointer/SPACE bookkeeping in lockstep.
- predefined movers: widen the unrolled-loop counter from unsigned long
  to size_t so a 64-bit element count cannot truncate on LLP64.
- test/datatype: make to_self --validate abort the job on a mismatch
  instead of exiting 0, so a validation failure is actually observable.
- gitignore: drop stale checksum and opal_crc entries left behind by
  the earlier checksum/CRC removal.

+++ opal/datatype/opal_datatype_pack_unpack_predefined.h
@@ -221,7 +221,7 @@ _Static_assert(OPAL_DATATYPE_PREDEFINED_MAX_INLINE_BLOCKLEN == 8,
 
 #define OPAL_DATATYPE_PACK_PREDEFINED_ELEMENT_UNCHECKED(src_base, dest_base, count, blocklen, type) \
     {                                                                                               \
-        register unsigned long i = count;                                                           \
+        register size_t i = count;                                                                  \
         type *_src = (type *) src_base;                                                             \
         type *_dest = (type *) dest_base;                                                           \
         if (0 == (elem->extent % (ptrdiff_t) sizeof(type))) {                                       \
@@ -265,7 +265,7 @@ _Static_assert(OPAL_DATATYPE_PREDEFINED_MAX_INLINE_BLOCKLEN == 8,
     {                                                                                                 \
         type *_src = (type *) src_base;                                                               \
         type *_dest = (type *) dest_base;                                                             \
-        register unsigned long i = count;                                                             \
+        register size_t i = count;                                                                    \
         if (0 == (elem->extent % (ptrdiff_t) sizeof(type))) {                                         \
             /* Mirror of pack: the packed source is contiguous and the destination strides by a      \
              * whole number of elements, so advance in element units for scaled-index addressing. */ \
